// packages/core/src/types.d.ts
export interface ChunkUploadOptions {
    chunkSize?: number;
    maxConcurrent?: number;
    autoRetry?: boolean;
    maxRetries?: number;
    headers?: Record<string, string>;
    withCredentials?: boolean;
    context?: string;
    checksum?: boolean;
    endpoints?: {
        initiate?: string;
        upload?: string;
        status?: string;
        cancel?: string;
    };
}